[ghstack-poisoned]

+++ torch/csrc/jit/passes/common_subexpression_elimination.cpp
@@ -1,14 +1,16 @@
 #include <torch/csrc/jit/passes/common_subexpression_elimination.h>
 
-#include <torch/csrc/jit/ir.h>
-#include <torch/csrc/jit/node_hashing.h>
-#include <torch/csrc/jit/passes/alias_analysis.h>
+#include <torch/csrc/jit/ir/alias_analysis.h>
+#include <torch/csrc/jit/ir/ir.h>
+#include <torch/csrc/jit/ir/node_hashing.h>
+#include <torch/csrc/jit/jit_log.h>
 
 #include <unordered_map>
 
 namespace torch {
 namespace jit {
 namespace {
+
 // The function implements common subexpression elimination.
 // Since the nodes are visited in topological order, one pass is enough.
 void EliminateCommonSubexpression(
@@ -18,8 +20,16 @@ void EliminateCommonSubexpression(
   std::unordered_set<Node*, HashNode, EqualNode> subexprs;
   for (auto it = block->nodes().begin(); it != block->nodes().end(); ++it) {
     auto node = *it;
-    if (node->hasSideEffects() || node->isNondeterministic() ||
-        aliasDb.hasWriters(node)) {
+    if (node->hasSideEffects()) {
+      GRAPH_DEBUG("Node was skipped due to side effects:\n", *node);
+      continue;
+    }
+    if (node->isNondeterministic()) {
+      GRAPH_DEBUG("Node was skipped due to its non determinism:\n", *node);
+      continue;
+    }
+    if (aliasDb.hasWriters(node)) {
+      GRAPH_DEBUG("Node was skipped due to alias analysis result:\n", *node);
       // Do NOT have enough information to do CSE on these nodes.
       continue;
     }
@@ -42,7 +52,14 @@ void EliminateCommonSubexpression(
 
     // Check for CSE opportunities in the parent block.
     auto parent_lookup = parent_lookup_fn(node);
+    auto g_out = node->owningGraph()->outputs();
     if (parent_lookup) {
+      if (!aliasDb.safeToChangeAliasingRelationship(
+              node->outputs(), parent_lookup->outputs())) {
+        continue;
+      }
+
+      GRAPH_UPDATE("Replacing\n", *node, "with\n", *parent_lookup);
       node->replaceAllUsesWith(parent_lookup);
       it.destroyCurrent();
       continue;
@@ -53,6 +70,15 @@ void EliminateCommonSubexpression(
     if (!subit.second) {
       // Subexpression exists, replace the uses of node, and destroy it.
       auto existing = *subit.first;
+
+      // don't introduce new aliasing among graph outputs
+      if (aliasDb.mayContainAlias(
+              node->outputs(), node->owningGraph()->outputs()) &&
+          aliasDb.mayContainAlias(existing->outputs(), g_out)) {
+        continue;
+      }
+
+      GRAPH_UPDATE("Replacing\n", *node, "with\n", *existing);
       node->replaceAllUsesWith(existing);
       // Destroy the node.
       it.destroyCurrent();
@@ -61,8 +87,9 @@ void EliminateCommonSubexpression(
 }
 } // namespace
 
-void EliminateCommonSubexpression(std::shared_ptr<Graph>& graph) {
+void EliminateCommonSubexpression(const std::shared_ptr<Graph>& graph) {
   AliasDb aliasDb(graph);
+  GRAPH_DUMP("Before CSE", graph);
   EliminateCommonSubexpression(
       graph->block(), aliasDb, [](Node*) { return nullptr; });
 }
